+++ libweva/tests/main.cpp
@@ -540,6 +540,7 @@ void test_abi_tab_navigation();
 void test_abi_embedded_focus_and_text();
 void test_abi_key_repeat();
 void test_abi_double_click();
+void test_abi_text_properties();
 void test_abi_keyboard_activation();
 void test_abi_ime_composition();
 void test_abi_ime_external_mutation();
@@ -1122,6 +1123,7 @@ int main() {
     test_abi_embedded_focus_and_text();
     test_abi_key_repeat();
     test_abi_double_click();
+    test_abi_text_properties();
     test_abi_keyboard_activation();
     test_abi_ime_composition();
     test_abi_ime_external_mutation();